backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.8.3",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pymongo[srv]>=4.14.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scipy>=1.16.1",
    "spacy>=3.8.7",
    "uvicorn>=0.35.0",
]
